+++ PJ_RESTAURANTE_5TO_CICLO/restaurante/app/src/main/AndroidManifest.xml
@@ -13,7 +13,13 @@
         android:theme="@style/Theme.Restaurante"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.perfilUsuario.EditarPerfilActivity"
+            android:name=".presentation.Perfil.MisDirecciones.MisDireccionesActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.Perfil.CambiarContrasena.CambiarContrasenaActivity"
+            android:exported="false" />
+        <activity
+            android:name=".presentation.Perfil.EditarPerfil.EditarPerfilActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.home.HomeActivity"
@@ -22,7 +28,7 @@
             android:name=".presentation.StartActivity"
             android:exported="false" />
         <activity
-            android:name=".presentation.perfilUsuario.PerfilUsuarioActivity"
+            android:name=".presentation.Perfil.PerfilUsuarioActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.registro.RegistroActivity"
