--- PJ_RESTAURANTE_5TO_CICLO/restaurante/app/src/main/java/com/example/restaurante/presentation/perfilUsuario/PerfilUsuarioActivity.kt
@@ -1,15 +1,0 @@
-package com.example.restaurante.presentation.perfilUsuario
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import com.example.restaurante.databinding.ActivityPerfilUsuarioBinding
-
-class PerfilUsuarioActivity : AppCompatActivity() {
-    private lateinit var binding : ActivityPerfilUsuarioBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityPerfilUsuarioBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-}
