+++ PJ_RESTAURANTE_5TO_CICLO/restaurante/app/src/main/res/layout/activity_mis_direcciones.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".presentation.Perfil.MisDirecciones.MisDireccionesActivity">
+
+</androidx.constraintlayout.widget.ConstraintLayout>
